+++ tut24 gbsound1/main.c
@@ -80,7 +80,7 @@ int main(void)
 	ST_PARAM param[8];
 
 	bool isSnd = false;
-	s16 sel = 0;
+	s32 sel = 0;
 
 	_Memset(&param, 0x00, sizeof(param));
 
@@ -144,8 +144,8 @@ int main(void)
 	    VBlankIntrWait();
 
 		KeyExec();
-		u16 rep = KeyGetRep();
-		u16 trg = KeyGetTrg();
+		u32 rep = KeyGetRep();
+		u32 trg = KeyGetTrg();
 
 		if(rep & KEY_UP && sel >  0)
 		{
