+++ tut27 gbsound4/main.c
@@ -55,7 +55,7 @@ int main(void)
 	ST_PARAM param[9];
 
 	bool isSnd = false;
-	s16 sel = 0;
+	s32 sel = 0;
 
 	_Memset(&param, 0x00, sizeof(param));
 
@@ -125,8 +125,8 @@ int main(void)
 	    VBlankIntrWait();
 
 		KeyExec();
-		u16 rep = KeyGetRep();
-		u16 trg = KeyGetTrg();
+		u32 rep = KeyGetRep();
+		u32 trg = KeyGetTrg();
 
 		if(rep & KEY_UP && sel >  0)
 		{
